Move transformer functions to new utils file, implement fix_combination_order function
